feat(web): GitHub settings page polish — section headings, two-step panel, Source Repos move
UX pass on Settings → GitHub after the connect-panel ship:

- The page now has two titled sections: "GitHub Connection" (the
  installation panel) and "Source Repos" — the repo runtime resources,
  moved here from Settings → Resources so the repos agents work on sit
  next to the GitHub connection their code flows through. The resource
  section machinery is extracted into a shared ResourceTypeSections
  component; Resources keeps prompts / skills / MCPs.
- Summary card: "Connected as" → "Connected to", GitHub accounts render
  as github.com/<login> everywhere (never confusable with a First Tree
  team name), and the collapsed Connection details sit directly under
  the connected account instead of below the action buttons.
- The connect panel reads as the real flow's two steps: "Step 1:
  Install the First Tree App on your GitHub" (or, once connected,
  "GitHub App installed." with Reinstall + Manage on GitHub) and
  "Step 2: Connect to your GitHub" with the trimmed waiting line
  "Waiting for GitHub… You may need a GitHub org admin to approve."
