+++ packages/web/src/pages/__tests__/resource-editors-dom.test.tsx
@@ -107,7 +107,12 @@ async function clickWhenEnabled(text: string): Promise<void> {
 }
 
 async function render(): Promise<{ root: Root }> {
-  const { SettingsResourcesPage } = await import("../settings/resources.js");
+  // Mount the shared sections component with every type: the machinery under
+  // test is type-agnostic, and the page-level split (repo on Settings →
+  // GitHub as "Source Repos", the rest on Settings → Resources) is covered
+  // by the page smoke tests.
+  const { ResourceTypeSections } = await import("../settings/resource-sections.js");
+  const { RESOURCE_TYPES } = await import("../settings/resource-editors.js");
   const { ToastProvider } = await import("../../components/ui/toast.js");
   const container = document.createElement("div");
   document.body.appendChild(container);
@@ -116,7 +121,7 @@ async function render(): Promise<{ root: Root }> {
     root.render(
       <QueryClientProvider client={createClient()}>
         <ToastProvider>
-          <SettingsResourcesPage />
+          <ResourceTypeSections types={RESOURCE_TYPES} />
         </ToastProvider>
       </QueryClientProvider>,
     );
